Cater for strings and objects when filtering upload component steps
Both the `includeUploadSteps` and `excludeUploadSteps` functions
are used to filter upload component steps. The parameter passed
to them can either be an object or a string depending on where
they are invoked.

We had previously adjusted these functions to fix the reordering of
components by making sure an actual ID is passed to `excludeUploadSteps`
However `step` at that moment is not guaranteed to be an object.

Here we allow for the parameters to be either.